+++ changelog.md
@@ -1,5 +1,9 @@
 # ACEmulator Change Log
 
+[Og II]
+* Added the ability to equip items and have it also update your visual appearance.
+* We have some data work to do - to add clothing table did values, but we have added 30 or so for testing out of 150+ needed.
+
 ### 2017-06-23
 [ddevec]
 * Fix to teleport freezing characters in place (Wasn't clearing requested location)
@@ -39,6 +43,7 @@
 * Please call DbManager.SaveObject() now instead of DatabaseManager.Shard.SaveObject()
 
 ### 2017-06-19
+[Og II]
 * Trello card task - Extend ORM code to support multiple getLists or implement another method to query by alternate keys - re-implement cirand
 * Many thanks to ddevec - he did the work on the ORM code revision.   I just tested and implemented for my use.
 * Fixed undiscovered alignment bug at hooktype we were sending as dword and it is a word.   That was shifting everything below it.